Add Sprot related errors (#82)
This adds error types that can be returned by sprot. Each of the errors is actually from a distinct hubris API, but all are combined into the top-level SprotError. This is supported in hubris via the new Complex errors in idol. A bunch of From implementations exist in oxidecomputer/hubris#1279 to convert the underlying hubris errors to the gateway-messages errors.

This PR puts sprot related errors into both SpError and RotError. It's unclear if we actually want to maintain RotError at all. Currently it's only used in when getting state from the RoT.

The errors in hubris map 1:1 to the gateway-messages errors right now, with the addition of a currently unused Unknown(u32) variant in each gateway-messages error. This is there to support upgrades. The plan for upgrade
is that new variants will be added to the hubris error and the From implementation in hubris will put the index of the new error into the Unknown variant in the gateway-messages version. This gives us a way to figure out what error occurred by looking at the from implementation. This allows updated SPs and RoTs to work with older MGSs. This situation, however is expected to be only temporary. The new variant will be added to the gateway-messages type and then MGSs will be upgraded. Once gateway-messages has the new variant, hubris will be updated to pull in the newer version of gateway-messages and the From implementation will be updated, and the Unknown mapping will be removed as it is no longer needed.

We also plan to ensure that MGS can talk to older versions of the SP, and so if we can update MGS first we will. Then multiple updates will not be required to get MGS and the SP in sync. In this case we'd just add the new variants to MGS and hubris simultaneously and never use the Unknown gateway-messages variant.

In all cases, there needs to be an update mechanism capable of making these decisions about what is possible to update and what is not.