Add `data_sent` to `HostPhase2Progress` reports
This required some refactoring to send the progress reports _after_
we've requested data from the host phase2 provider and sent data to the
SP, which probably makes them more accurate anyway! Previously we would
update our progress even if the request from the SP was nonsense or we
failed to reply to it; now we only update progress after we know we've
successfully sent data to an SP (although of course we don't know for
sure that it received it).