Add data sent to host phase2 progress (#87)
This required some refactoring to send the progress reports after
we've requested data from the host phase2 provider and sent data to the
SP, which probably makes them more accurate anyway! Previously we would
update our progress even if the request from the SP was nonsense or we
failed to reply to it; now we only update progress after we know we've
successfully sent data to an SP (although of course we don't know for
sure that it received it).

The primary motivator for this is correctness in wicket: prior to this PR, progress
for delivery of the trampoline image looks like it stalls at 99.9%, because we
were reporting the start of the last data request from the SP, but we really
want to report the end. That requires the changes in this PR: we need to know
how much data we sent in response to the SP's request.