Use a higher RPC timeout count when resetting the SP
`SingleSp` allows its clients to choose a max retry count and delay
between retries, and most of our clients currently choose values that
result in an overall timeout of 3-5 seconds. This is more than enough
for _nearly_ every request the SP can service, but is definitely not
enough when asking the SP to reset itself (particularly if the SP in
question is a sidecar SP, in which case it can't respond until it brings
the management network back online).

This change internally bumps the retry count up sufficiently to allow
for ~30 seconds when resetting an SP, while still keeping the
client-requested timeout for all other requests.