[gateway-messages] Introduce SpStateV2 (#92)
* [gateway-messages] Introduce SpStateV2

We introduce a new response for SpState requests. By modifying the
expected response in gateway-sp-comms, we can receive both the original
`SpState` and `SpStateV2` responses enclosed in a new `VersionedSpState`
response. Faux-mgs just prints this, but MGS can return the combined
results and allow upstream callers to decide what to do. This is a
general strategy for message evolution, and in some cases MGS can only
return the latest version by filling in optional fields with `None` if a
new field is the only thing added to the new version of the struct.

Additionally, the strategy used here makes things easy on control-plane-
agent as it only has to return the latest version of the type that it
knows about.