[faux-mgs] Make console less panic-happy (#124)
All the `.unwrap()`s have been replaced. Most of them were fine and
became `.expect("reason we think it's fine")`. The one that was not is
in #122: if we fail a send to the SP, we previously panicked. We now
buffer failed sends and will retry them (indefinitely, logging on each
failure).