Relax SP slot update requirements
The SP has dual bank flash. The model in almost all of the system
is that bank 0 (slot 0) is the active running image and bank 1
(slot 1) is the inactive slot. The 'almost' is that we require the
slot for update to be 0. This doesn't actually match with what
we're actually doing so change the error to a warning if we aren't
passing in slot 1 to allow clients to better match what is happening.

While we're at it, make this detail of the SP part of our MGS
protocol