Add `MeasurementKind::Dimensionless`
This commit adds a `Dimensionless` variant to the `MeasurementKind` enum
to represent measurements without units. This is intended to be used to
represent AMD CPU T<sub>ctl</sub> thermal values, which are a unitless
value from 0-100 representing a kind of "abstract thermal throttling
danger level" --- see oxidecomputer/hubris#1881 for details on why we
need to differentiate this from other temperature measurements.

I wasn't totally sure as to the best approach for naming this variant. I
note that the name of the enum is "measurement kind", not "measurement
unit", so it doesn't directly represent the unit of the measurement but
rather, the kind of quantity being measured. So, perhaps this ought to
be a `CpuTctl` variant or similar, to *specifically* represent
T<sub>ctl</sub> measurements. If other dimensionless values show up in
future, they would then have their own variant. I'm open to being
convicned either way --- what do others think?