ereport-messages: Remove COMMIT flag from requests (#388)
As discussed in [this comment][1] from @cbiffle, the `C`/`COMMIT` flag
in ereport request messages is unnecessary, as ENA 0 is reserved by the
SP to indicate "no ENA". So, if we want to say "no ENAs have been
committed", MGS can just send ENA 0. This makes the request message an
entire byte smaller (wow! whoa!), and also lets us remove the dependency
on the `bitflags` crate, which in turn means less code in `packrat`.

I've changed the name of the `Ena::ZERO` constant to `Ena::NONE` to
communicate that ENA 0 means "no ENA".

[1]: https://github.com/oxidecomputer/rfd/pull/888/files/7146ee385715d1ddc6b7dc2462df7da80ec3bde7#r2078301646