Add `Debug` logging for  power state transitions (#391)
While working on integrating the changes from #390 in Omicron, I
realized that I should really have added a `Debug` implementation for
`PowerStateTransition`, so that MGS can log the outcome of a
`set_power_state` request without having to have two separate log macro
invocations in a branch. This commit adds `#[derive(Debug)]` to
`PowerStateTransition` in service of that.

Furthermore, `faux-mgs` should probably indicate to the user whether a
request to set the power state results in an actual change in state, so
I've updated it to do that. Whether or not a transition occurred is now
included in the `slog` logging, textual output, and JSON output from
`faux-mgs`.