Implement ComponentCancelPendingActiveSlot
This message can be used in favor of resetting the RoT when recovering
from a failed or abandoned update.

A RoT transient or persistent Hubris activation has a pending phase
that is resolved on next boot.

Clearing the pending activation allows a new update flow to proceed
without resetting the RoT, or with a reset, allows the update flow to
proceed without switching to a different image (assuming a properly
signed alternate image).