gateway-sp-comms: fix handling of empty metadata maps
Presently, the `ereport::Worker` struct [stores the metadata map in an
`Option`][1]. Metadata refresh requests (`restart_id=0, start_ena=0,
limit=0`) are [sent to the SP if the `Option` is `None`][2]. The option
[is set to `Some`][3] if we receive a packet from the SP where the
metadata map is non-empty, or if the restart ID mismatches the requested
one.

If I recall correctly, the `Option` was intended to distinguish between
"we just started up" and "we received an explicit empty metadata map".
But, I don't actually think we _should_ be distinguishing between those
cases. When the SP has restarted and given us an empty metadata map,
this may be because we requested ereports from `packrat` _before_ VPD
has been loaded (as I discussed in
https://github.com/oxidecomputer/hubris/pull/2126#discussion_r2193653564).
In that case, when the SP sends us an empty metadata map, we want to
keep requesting the metadata on every subsequent request, as it might be
set later.

Thus, this commit just removes the `Option` and has it start out with an
_empty_ map, and overwrites the existing map if the restart IDs are
mismatched, *or* any time the current map is empty and the received one
is non-empty. I've also added an additional test for this behavior.

Fixes #409

[1]:
https://github.com/oxidecomputer/management-gateway-service/blob/77e316c812aa057b9714d0d99c4a7bdd36d45be2/gateway-sp-comms/src/ereport.rs#L79-L83
[2]:
https://github.com/oxidecomputer/management-gateway-service/blob/77e316c812aa057b9714d0d99c4a7bdd36d45be2/gateway-sp-comms/src/ereport.rs#L109-L111
[3]:
https://github.com/oxidecomputer/management-gateway-service/blob/77e316c812aa057b9714d0d99c4a7bdd36d45be2/gateway-sp-comms/src/ereport.rs#L351-L360