Improve ignition target selection in `faux-mgs` (#467)
Ignition commands in faux-mgs require specifying an "ignition target", which has
to be looked up in RFD 144 § Switch port map. This isn't the same as cubby
number, which is annoying.

This PR adds `--target`, `--cubby`, `--sidecar`, and `--psc` selector arguments
to specify an ignition target at a high level.

The positional argument is still supported when possible (`ignition` and
`ignition-link-events`), but is removed from `clear-ignition-link-events` and
`ignition-command` because of Clap limitations.